docs(demos): convert inject-form to folder pattern + drop `!` teaching
The inject-form demo had two inline render-function components
(defineComponent + h) that taught a discouraged pattern. Splits the
children into their own SFCs alongside a shared schema.ts, deletes the
shadowed flat use-register.vue (the use-register folder demo has been
the live path since c184392), and teaches the demos index + playground
to surface folder demos with arbitrary sibling files (.vue / .ts / .js
/ .css), so `import { schema } from './schema'` keeps resolving inside
the REPL.

Also drops the `!` non-null assertion from `injectForm` calls in the
new demos and the docs page that taught it. Children keep
`ctx: T | null` and reach for fields via `?.` at every consumption
site, so a mount-order regression degrades to a quiet no-op instead of
crashing the subtree.
